chore(arch): move refreshScheduler to features/auth + envelope ADR
  - Move shared/lib/auth/refreshScheduler.{ts,test.ts} to features/auth/lib/.
    Fixes the only shared/ → features/ import (boundary rule #2).
  - Update useAuth.ts import path.
  - docs/architecture.md: fix scheduler path + add ADR for the
    { data, meta } response envelope (typed raw in api/client, unwrapped in hooks).
